+++ robotframework-ls/tests/robotframework_ls_tests/_resources/atest_v5/variables/resvarfiles/cli_vars_2.py
@@ -1,0 +1,10 @@
+def get_variables(name, value='default value'):
+    if name == 'FAIL':
+        1/0
+    varz = { name: value,
+             'ANOTHER_SCALAR': 'Variable from CLI var file with get_variables',
+             'LIST__ANOTHER_LIST': ['List variable from CLI var file',
+                                    'with get_variables'] }
+    for name in 'PRIORITIES_1', 'PRIORITIES_2', 'PRIORITIES_2B':
+        varz[name] = 'Second Variable File from CLI'
+    return varz