+++ robotframework-ls/tests/robotframework_ls_tests/_resources/atest_v5/variables/same_variable_file_names/different_variable_files/suite1/subsuite1/variable.py
@@ -1,0 +1,1 @@
+SUITE = SUITE_11 = "suite1.subsuite1"