+++ robotframework-ls/tests/robotframework_ls_tests/_resources/atest_v5/variables/same_variable_file_names/different_variable_files/suite3/subsuite1/variable.py
@@ -1,0 +1,1 @@
+SUITE = SUITE_31 = "suite3.subsuite1"