+++ robotframework-ls/tests/robotframework_ls_tests/_resources/atest_v5/variables/same_variable_file_names/different_variable_files/suite3/variable.py
@@ -1,0 +1,1 @@
+SUITE = SUITE_3 = "suite3"