+++ robotframework-ls/tests/robotframework_ls_tests/_resources/atest_v5/variables/scalar_lists.py
@@ -1,0 +1,17 @@
+LIST = ['spam', 'eggs', 21]
+
+
+class _Extended:
+    list = LIST
+    string = 'not a list'
+    def __getitem__(self, item):
+        return LIST
+
+EXTENDED = _Extended()
+
+
+class _Iterable:
+    def __iter__(self):
+        return iter(LIST)
+
+ITERABLE = _Iterable()